+++ packages/observatory/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ue-shed/observatory
 
+## 0.4.0
+
+### Patch Changes
+
+- @ue-shed/observability@0.4.0
+    - @ue-shed/unreal-connection@0.4.0
+
 ## 0.3.0
 
 ### Patch Changes