+++ packages/project-custodian/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ue-shed/project-custodian
 
+## 0.4.0
+
 ## 0.3.0
 
 ## 0.2.0