+++ packages/protocol/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ue-shed/protocol
 
+## 0.4.0
+
 ## 0.3.0
 
 ### Minor Changes