+++ packages/uasset-inspection-wasm/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ue-shed/uasset-inspection-wasm
 
+## 0.4.0
+
 ## 0.3.0
 
 ## 0.2.0