+++ packages/unreal-connection/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ue-shed/unreal-connection
 
+## 0.4.0
+
+### Patch Changes
+
+- @ue-shed/protocol@0.4.0
+
 ## 0.3.0
 
 ### Patch Changes